update on sendMail
